Tutorial #11

+++ src/main/java/utils/Time.java
@@ -2,7 +2,7 @@ package utils;
 
 public class Time {
 
-	public static float timeStarted = System.nanoTime();
+	public static long timeStarted = System.nanoTime();
 
 	public static float getTime() {
 		return (float) ((System.nanoTime() - timeStarted) * 1E-9);
